% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-alpha_test.R
\name{alpha_test,single_period_at_data,at_settings_factor_z-method}
\alias{alpha_test,single_period_at_data,at_settings_factor_z-method}
\title{Alpha Test Method for Single Period AT}
\usage{
\S4method{alpha_test}{single_period_at_data,at_settings_factor_z}(data, .settings, ...)
}
\description{
Alpha Test Method for Single Period AT
}
