% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{alpha_test}
\alias{alpha_test}
\title{Alpha Test}
\usage{
alpha_test(data, .settings, ...)
}
\arguments{
\item{data}{factor_data or single_period_factor_data object}

\item{.settings}{factor_data_params object}
}
\value{
alpha_test or single_period_at object
}
\description{
Run Alpha Test on factor data with at_settings
}
