% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_test_z.R
\docType{class}
\name{alpha_test_z-class}
\alias{alpha_test_z-class}
\title{Alpha Test (Factor Z-Score) S4 Object}
\description{
Alpha Test (Factor Z-Score) S4 Object
}
\section{Slots}{

\describe{
\item{\code{IC}}{Information Coefficient}

\item{\code{factor_z_score}}{A list of factor z-scores.}
}}

