% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-at_data.R
\docType{class}
\name{at_data-class}
\alias{at_data-class}
\title{Alpha Test Data (S4 Object)}
\description{
An S4 class to represent alpha test data for multiple periods.
}
\section{Slots}{

\describe{
\item{\code{alpha_data}}{A list of `single_period_at_data` objects representing
alpha test results for individual periods.}

\item{\code{data}}{A `data.frame` containing the raw input data used in the
alpha test.}

\item{\code{params}}{An `at_data_params` object containing the parameters used
to configure the alpha test.}
}}

