% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-at_data_params.R
\docType{class}
\name{at_data_params-class}
\alias{at_data_params-class}
\title{Alpha Test Data Parameters}
\description{
An S4 class containing the parameters for creating a at_data object
}
\section{Slots}{

\describe{
\item{\code{data_col_name}}{A string representing the column name of date values.}

\item{\code{id_col_name}}{A string representing the column name of id values.}

\item{\code{factor_col_name}}{A string representing the column name of factor values.}

\item{\code{return_col_name}}{A string representing the column name of return values.}

\item{\code{include_col_name}}{A string representing the column name of include T/F.}

\item{\code{group_col_name}}{A string representing the column name of grouping value.}

\item{\code{weight_col_name}}{A string representing the column name of weights.}

\item{\code{horizon}}{return and factor horizon}
}}

