% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATSettings.R
\name{AT_FactorWeighted}
\alias{AT_FactorWeighted}
\title{Factor Weighted Alpha Testing Settings}
\usage{
AT_FactorWeighted(win.prob = c(0, 1), longonly = F, ...)
}
\arguments{
\item{win.prob}{A numeric vector of length 2 representing the percentile
cut-offs for windsorization}

\item{longonly}{A logical representing whether or not the portfolio weights
should be long only (True) or Long-Short (False)}
}
\value{
An ATSettings_FactorWeighted S4 object to be used in Alpha Testing.
}
\description{
A function to create Alpha Testing settings for a factor weighted Alpha Test.
}
