% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATSettings.R
\name{AT_SettingQSpread}
\alias{AT_SettingQSpread}
\title{Quantile Spread Alpha Testing Settings}
\usage{
AT_SettingQSpread(
  start.date = as.Date("1901-01-01"),
  end.date = Sys.Date(),
  quantiles = 5,
  longonly = F,
  leverage = 1
)
}
\arguments{
\item{start.date}{description}

\item{end.date}{description}

\item{quantiles}{description}

\item{longonly}{A logical representing whether or not the portfolio weights
should be long only (True) or Long-Short (False)}

\item{leverage}{description}
}
\value{
An ATSettings_FactorWeighted S4 object to be used in Alpha Testing.
}
\description{
A function to create Alpha Testing settings for a Quantile Spread Alpha Test.
}
