% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_settings.R
\docType{class}
\name{at_settings_factor_z-class}
\alias{at_settings_factor_z-class}
\title{Settings for Factor Zscore Weighted Alpha Testing (S4 Object)}
\description{
Settings for Factor Zscore Weighted Alpha Testing (S4 Object)
}
\section{Slots}{

\describe{
\item{\code{win_prob}}{A numeric vector of length 2 representing the percentile
cut-offs for windsorization.}
}}

