% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATSettings.R
\name{AT_SettingWeighted}
\alias{AT_SettingWeighted}
\title{Factor Weighted Alpha Testing Settings}
\usage{
AT_SettingWeighted(
  start.date = as.Date("1901-01-01"),
  end.date = Sys.Date(),
  win.prob = c(0, 1),
  longonly = F,
  leverage = 1
)
}
\arguments{
\item{start.date}{description}

\item{end.date}{description}

\item{win.prob}{A numeric vector of length 2 representing the percentile
cut-offs for windsorization.}

\item{longonly}{A logical representing whether or not the portfolio weights
should be long only (True) or Long-Short (False)}

\item{leverage}{description}
}
\value{
An ATSettings_FactorWeighted S4 object to be used in Alpha Testing.
}
\description{
A function to create Alpha Testing settings for a factor weighted Alpha Test.
}
