% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATSettings.R
\docType{class}
\name{ATSettings-class}
\alias{ATSettings-class}
\title{Alpha Testing Settings}
\description{
Alpha Testing Settings
}
\section{Slots}{

\describe{
\item{\code{WeightingScheme}}{Character}
}}

