% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATSettings.R
\name{ATSettings}
\alias{ATSettings}
\title{Create Alpha Testing Settings}
\usage{
ATSettings(
  start.date = as.Date("1901-01-01"),
  end.date = Sys.Date(),
  testing.scheme = "Factor",
  ...
)
}
\arguments{
\item{start.date}{A date object representing the earliest start date to be
used when performing alpha testing. Factor data is filtered to start at or
after provided Start.Date argument.}

\item{end.date}{A date object representing the last date to be
used when performing alpha testing. Factor data is filtered to end at or
before provided End.Date argument.}

\item{testing.scheme}{Weighting scheme to be used in testing factor. 
Currently supports "Factor" and "Quantile"}

\item{...}{Additional settings to be passed based on Weighting.Scheme}
}
\description{
This function creates an "ATSettings" class with arguments to be used when
testing factors
}
