% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATSettings.R
\docType{class}
\name{ATSettings_FactorWeighted-class}
\alias{ATSettings_FactorWeighted-class}
\title{Settings for Factor Weighted Alpha Testing}
\description{
Settings for Factor Weighted Alpha Testing
}
\section{Slots}{

\describe{
\item{\code{win.prob}}{A numeric vector of length 2 representing the percentile
cut-offs for windsorization.}

\item{\code{longonly}}{A logical representing if portfolio weights should be long only.}

\item{\code{leverage}}{description}
}}

