% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATSettings.R
\docType{class}
\name{ATSettings_QSpread-class}
\alias{ATSettings_QSpread-class}
\title{Settings for Quantile Spread Alpha Testing}
\description{
Settings for Quantile Spread Alpha Testing
}
\section{Slots}{

\describe{
\item{\code{quantiles}}{Number of Quantiles to use.}

\item{\code{longonly}}{A logical representing if portfolio weights should be long only.}

\item{\code{leverage}}{A numeric representing the leverage.}
}}

