% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATSettings.R
\docType{class}
\name{ATSettings_Quantile-class}
\alias{ATSettings_Quantile-class}
\title{Settings for Quantile Alpha Testing}
\description{
Settings for Quantile Alpha Testing
}
\section{Slots}{

\describe{
\item{\code{quantiles}}{Number of Quantiles to use.}

\item{\code{longonly}}{A logical representing if portfolio weights should be 
long only}
}}

