% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-calc_bench_weights.R
\name{calc_bench_weights}
\alias{calc_bench_weights}
\title{Calculate Benchmark Weight for Alpha Testing}
\usage{
calc_bench_weights(bench_weighting_scheme = "zero", weights)
}
\arguments{
\item{bench_weighting_scheme}{Character representing the weighting scheme
to be applied. Possible values include "zero", "weighted", "equal-weighted",
"short-weighted", and "short-equal-weighted"}

\item{weights}{Numeric array of given weights.}
}
\description{
Function to calculate the weights for the benchmark returns.
}
