% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-calc_factor_q.R
\name{calc_factor_q,single_period_at_data-method}
\alias{calc_factor_q,single_period_at_data-method}
\title{Calculate Quantile-Based Factor Values}
\usage{
\S4method{calc_factor_q}{single_period_at_data}(at_data, quantiles = 5, .desc = TRUE)
}
\arguments{
\item{at_data}{A `single_period_at_data` object containing factor data.}

\item{quantiles}{A numeric value representing the number of quantiles.}

\item{.desc}{A logical value indicating whether values should be ranked in
descending order.}
}
\value{
An `orderedList` object containing `factor_q_score` objects.
}
\description{
Computes quantile-based factor scores for a given period using specified
quantiles and settings.
}
