% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_q_score.R
\name{calc_factor_q.single_period_at_data}
\alias{calc_factor_q.single_period_at_data}
\title{Calculate Quantile of Factor Values}
\usage{
\method{calc_factor_q}{single_period_at_data}(x, quantiles = 5, .desc = TRUE)
}
\arguments{
\item{x}{A single_period_at_data S4 object.}

\item{quantiles}{a numeric representing the number of quantiles}
}
\value{
orderedList S4 object of factor_q_score S4 objects
}
\description{
Calculate the Quantile of Factor Values including lagged
 Factor Values in a single_period_at_data S4 object
}
