% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_z_score.R
\name{calc_factor_z}
\alias{calc_factor_z}
\title{Calculate Z-Score Factor Value}
\usage{
calc_factor_z(factor_data, win_prob = c(0, 1))
}
\arguments{
\item{win_prob}{a numeric vector of probabilities with values in [0,1]}

\item{factor_values}{Factor Values.}
}
\description{
Calculate Factor Z-Score
}
