% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-calc_return.R
\name{calc_return}
\alias{calc_return}
\title{Calculate Returns Based on Weights and Returns Data}
\usage{
calc_return(weights, returns, ...)
}
\arguments{
\item{weights}{A numeric vector or `orderedList` of portfolio weights.}

\item{returns}{A numeric vector or `orderedList` of asset returns.}
}
\value{
A numeric value or `orderedList` representing portfolio returns.
}
\description{
Computes the weighted return for portfolios using the given weights and
returns data.
}
