% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-calc_return_z.R
\name{calc_return_z,single_period_at_data-method}
\alias{calc_return_z,single_period_at_data-method}
\title{Compute Z-Scores of Returns in Single Period Data}
\usage{
\S4method{calc_return_z}{single_period_at_data}(at_data, win_prob = c(0, 1))
}
\arguments{
\item{at_data}{A `single_period_at_data` object containing returns data.}

\item{win_prob}{A numeric vector of probabilities used for winsorization.}
}
\value{
An `orderedList` of `return_z_score` objects.
}
\description{
Calculates Z-scores for return values in a single period alpha test dataset.
}
