% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-calc_return_z.R
\name{calc_return_z}
\alias{calc_return_z}
\title{Calculate Z-Score of Return Values}
\usage{
calc_return_z(at_data, win_prob = c(0, 1))
}
\arguments{
\item{at_data}{A single_period_at_data S4 object.}

\item{win_prob}{a numeric vector of probabilities with values in [0,1]
defining the windsorization levels.}
}
\description{
Calculate the Z-Score of Returns Values including forward
 Return Values in a single_period_at_data S4 object
}
