% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_z_score.R
\name{calc_return_z.single_period_at_data}
\alias{calc_return_z.single_period_at_data}
\title{Calculate Z-Score of Return Values}
\usage{
\method{calc_return_z}{single_period_at_data}(x, win_prob = c(0, 1))
}
\arguments{
\item{x}{A single_period_at_data S4 object.}

\item{win_prob}{a numeric vector of probabilities with values in [0,1]
defining the windsorization levels.}
}
\value{
orderedList S4 object of return_z_score S4 objects
}
\description{
Calculate the Z-Score of Returns Values including forward
 Return Values in a single_period_at_data S4 object
}
