% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-calc_weights.R
\name{calc_weights,factor_z_score-method}
\alias{calc_weights,factor_z_score-method}
\title{Calculate Portfolio Weights}
\usage{
\S4method{calc_weights}{factor_z_score}(score, weighting_scheme = "z-weighted", ...)
}
\arguments{
\item{score}{A factor score object (either `factor_z_score` or
`factor_q_score`).}

\item{weighting_scheme}{A string representing the weighting scheme to use.}
}
\value{
A numeric vector of weights.
}
\description{
Computes portfolio weights based on factor scores and a specified weighting
scheme.
}
