% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-calc_weights.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate Weight from Factor Score}
\usage{
calc_weights(score, weighting_scheme, ...)
}
\arguments{
\item{score}{A factor score object}

\item{weighting_scheme}{A string representing the weighting scheme to be
used to calculate the portfolio weights.}
}
\description{
Function to calculate the portfolio weights for a given scoring
and weighting scheme
}
