% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-create_at_data_params.R
\name{create_at_data_params}
\alias{create_at_data_params}
\title{Create Alpha Test Data Parameters Object}
\usage{
create_at_data_params(
  date_col_name = NA_character_,
  id_col_name = NA_character_,
  factor_col_name = NA_character_,
  return_col_name = NA_character_,
  include_col_name = NA_character_,
  group_col_name = NA_character_,
  weight_col_name = NA_character_,
  horizon = 12
)
}
\arguments{
\item{date_col_name}{A string representing the column name of date values.}

\item{id_col_name}{A string representing the column name of ID values.}

\item{factor_col_name}{A string representing the column name of factor
values.}

\item{return_col_name}{A string representing the column name of return
values.}

\item{include_col_name}{A string representing the column name for include
flags (TRUE/FALSE).}

\item{group_col_name}{A string representing the column name for grouping
values.}

\item{weight_col_name}{A string representing the column name for weights.}

\item{horizon}{A numeric value representing the return and factor horizon.
Defaults to 12.}
}
\value{
An `at_data_params` S4 object.
}
\description{
Function to create and configure an `at_data_params` object, used to set
parameters for factor data in alpha testing.
}
