% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_data.R
\name{create_factor_data}
\alias{create_factor_data}
\title{Create Factor Data}
\usage{
create_factor_data(
  data,
  date_col_name = NA_character_,
  id_col_name = NA_character_,
  factor_col_name = NA_character_,
  return_col_name = NA_character_,
  group_col_name = NA_character_,
  weight_col_name = NA_character_
)
}
\arguments{
\item{data}{A data frame containing, id column, data column, factor column,
return column.}

\item{date_col_name}{A character representing the column name of the dates.}

\item{id_col_name}{A character representing the column name of the
identifiers.}

\item{factor_col_name}{A character representing the column name of the
factor.}

\item{return_col_name}{A character representing the column name of the
returns.}
}
\value{
A FactorData object.
}
\description{
Function to create factor data from data frame.
}
