% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-create_factor_data_params.R
\name{create_factor_data_params}
\alias{create_factor_data_params}
\title{Create Factor Data Parameters Object}
\usage{
create_factor_data_params(
  date_col_name = NA_character_,
  id_col_name = NA_character_,
  factor_col_name = NA_character_,
  return_col_name = NA_character_,
  group_col_name = NA_character_,
  weight_col_name = NA_character_
)
}
\arguments{
\item{date_col_name}{A string representing the column name of date values.}

\item{id_col_name}{A string representing the column name of ID values.}

\item{factor_col_name}{A string representing the column name of factor
values.}

\item{return_col_name}{A string representing the column name of return
values.}

\item{group_col_name}{A string representing the column name for grouping
values.}

\item{weight_col_name}{A string representing the column name for weights.}
}
\value{
A `factor_data_params` S4 object.
}
\description{
Function to create and configure a `factor_data_params` object, used for
specifying factor-related data parameters.
}
