% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-create_at_data.R
\name{create_single_period_at_data_list}
\alias{create_single_period_at_data_list}
\title{Create Single Period Alpha Test Data List}
\usage{
create_single_period_at_data_list(at_data)
}
\arguments{
\item{at_data}{factor_data_object}
}
\description{
Create Single Period Alpha Test Data List
}
\keyword{internal}
