% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_data_single_period.R
\name{create_single_period_factor_data}
\alias{create_single_period_factor_data}
\title{Create single_period_factor_data S4 Object}
\usage{
create_single_period_factor_data(
  data,
  date_col_name,
  id_col_name,
  factor_col_name,
  return_col_name,
  group_col_name,
  weight_col_name,
  ...
)
}
\arguments{
\item{data}{A data frame containing id columns, return column and factor
value column.}

\item{date_col_name}{A character representing the column name of dates.}

\item{id_col_name}{A character representing the column name of identifiers.}

\item{factor_col_name}{A character string representing the
column name of the factor.}

\item{return_col_name}{A character string representing the
column name of the returns.}

\item{group_col_name}{A character string representing the column name of
grouping varialble.}

\item{weight_col_name}{A character string representing the column name
of weighting variable.}
}
\value{
A SinglePeriodFactorData object.
}
\description{
Function to create SinglePeriodFactorData object
}
\details{
This function converts data into SinglePeriodFactorData object.
}
\keyword{internal}
