% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities_Scoring.R
\name{ctq}
\alias{ctq}
\title{Convert Data Array to Quantiles}
\usage{
ctq(x, fftile)
}
\arguments{
\item{x}{a numeric array to be quantiled}

\item{fftile}{integer number of fractiles to use in splitting data}
}
\value{
ftile
}
\description{
Function to calculate quantiles.
}
\details{
Used in Alpha Testing Functions
}
