% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_q_score.R
\name{custom_cut_interval}
\alias{custom_cut_interval}
\title{Discretise numeric data into categorical}
\usage{
custom_cut_interval(x, n, labels = NULL, right = TRUE)
}
\arguments{
\item{x}{numeric vector.}

\item{n}{number of intervals to create.}

\item{labels}{lables for the levels of the resulting category.}

\item{right}{logical, indicating if the intervals should be closed
on the right (and open on the left) or vice versa.}
}
\description{
Discretise numeric data into categorical
}
