% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_data_dt}
\alias{expand_data_dt}
\title{Expand Dataset for Factors and Returns}
\usage{
expand_data_dt(data, date_col_name, id_col_name, data_col_name)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{date_col_name}{A string specifying the column name for dates.}

\item{id_col_name}{A string specifying the column name for IDs.}

\item{data_col_name}{A string specifying the column name for data values.}
}
\value{
An expanded data frame with missing combinations filled with `NA`.
}
\description{
Expands the dataset to include all combinations of dates and IDs, ensuring
missing data is handled.
}
