% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_data.R
\docType{class}
\name{factor_data-class}
\alias{factor_data-class}
\title{Factor Data (S4 Object)}
\description{
An S4 Class to represent Factor Data over time
}
\section{Slots}{

\describe{
\item{\code{factor_data}}{A list of SinglePeriodFactorData objects.}

\item{\code{data}}{A data.frame containing raw data.}

\item{\code{params}}{A factor_data_params object}
}}

