% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-factor_q_score.R
\docType{class}
\name{factor_q_score-class}
\alias{factor_q_score-class}
\title{Factor Quantile Scores (S4 Object)}
\description{
Encapsulates quantile-based factor scores for a given period.
}
\section{Slots}{

\describe{
\item{\code{score}}{A named vector of factor scores for the quantile-based analysis.}
}}

