% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-factor_z_score.R
\docType{class}
\name{factor_z_score-class}
\alias{factor_z_score-class}
\title{Factor Z-Score (S4 Object)}
\description{
Represents Z-scores of factor values for a single period in alpha testing.
}
\section{Slots}{

\describe{
\item{\code{score}}{A numeric vector containing Z-scores for factors.}
}}

