% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorData.R
\docType{class}
\name{FactorData-class}
\alias{FactorData-class}
\title{An S4 Class to represent Factor Data over time}
\description{
An S4 Class to represent Factor Data over time
}
\section{Slots}{

\describe{
\item{\code{data}}{A list of SinglePeriodFactorData objects.}
}}

