% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorData.R
\name{FactorData}
\alias{FactorData}
\title{Create Factor Data}
\usage{
FactorData(data, dname, iname, fname, rname)
}
\arguments{
\item{data}{A data frame containing, id column, data column, factor column, return column.}

\item{dname}{A character representing the column name of the dates.}

\item{iname}{A character representing the column name of the identifiers.}

\item{fname}{A character representing the column name of the factor.}

\item{rname}{A character representing the column name of the returns.}
}
\value{
A FactorData object.
}
\description{
Function to create factor data from data frame.
}
\details{
Details
}
