% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_fwd_p_returns}
\alias{get_fwd_p_returns}
\title{Compute Forward Returns}
\usage{
get_fwd_p_returns(data, date_col_name, id_col_name, return_col_name, horizon)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{date_col_name}{A string specifying the column name for dates.}

\item{id_col_name}{A string specifying the column name for IDs.}

\item{return_col_name}{A string specifying the column name for returns.}

\item{horizon}{An integer representing the forward horizon in periods.}
}
\value{
A data frame containing the computed forward returns.
}
\description{
Computes forward periodic returns for a given dataset based on specified
time horizons.
}
