% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_prv_factor_vals}
\alias{get_prv_factor_vals}
\title{Get lagged factor values}
\usage{
get_prv_factor_vals(data, date_col_name, id_col_name, factor_col_name, horizon)
}
\arguments{
\item{data}{Data frame}

\item{date_col_name}{string representing name of date column in data.}

\item{id_col_name}{string representing name of id column in data.}

\item{factor_col_name}{string representing name of factor column in data.}

\item{horizon}{numeric representing the number of periods to lag factor data.}
}
\description{
For every date, id get the factor and lagged values of factor.
}
