% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-multi_period_at.R
\docType{class}
\name{multi_period_at-class}
\alias{multi_period_at-class}
\title{Multi Period Alpha Testing S4 Object}
\description{
An S4 Class to represent Factor Alpha Testing Parent Calss
 for multiple periods.
}
\section{Slots}{

\describe{
\item{\code{return_fwd_factor}}{ToDo}

\item{\code{return_lag_factor}}{ToDo}

\item{\code{return_bmark}}{ToDo}

\item{\code{weights}}{ToDo}

\item{\code{weights_bmark}}{ToDo}

\item{\code{alpha_fwd_return}}{ToDo}

\item{\code{alpha_lag_factor}}{ToDo}
}}

