% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-multi_period_at_q.R
\docType{class}
\name{multi_period_at_q-class}
\alias{multi_period_at_q-class}
\title{Multi Period Alpha Test (Factor Quantile) S4 Object}
\description{
Multi Period Alpha Test (Factor Quantile) S4 Object
}
\section{Slots}{

\describe{
\item{\code{factor_quantile}}{A list of factor quantiles.}

\item{\code{q_returns}}{A list of factor quantile returns.}

\item{\code{q_stats}}{A list of factor quantile stats.}
}}

