% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-multi_period_at_z.R
\docType{class}
\name{multi_period_at_z-class}
\alias{multi_period_at_z-class}
\title{Multi Period Alpha Test (Factor Z-Score) S4 Object}
\description{
Multi Period Alpha Test (Factor Z-Score) S4 Object
}
\section{Slots}{

\describe{
\item{\code{IC_fwd_return}}{ToDo}

\item{\code{IC_lag_factor}}{ToDo}
}}

