% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-orderedList.R
\docType{class}
\name{orderedList-class}
\alias{orderedList-class}
\title{Ordered List (S4 Object)}
\description{
Represents a list with a defined order, often used in alpha testing
configurations.
}
\section{Slots}{

\describe{
\item{\code{list}}{A list containing the elements.}

\item{\code{n}}{An integer representing the length of the list.}

\item{\code{order}}{A numeric array indicating the order of elements in the list.}
}}

