% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderedList.R
\name{orderedList}
\alias{orderedList}
\title{Create orderedList}
\usage{
orderedList(list, n, order)
}
\arguments{
\item{list}{a list}

\item{n}{a numeric representing the length of the list}

\item{order}{a numeric array representing the order of the list}
}
\description{
Creates an orderedList S4 Object
}
