% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderedList.R
\name{pivot_lospat_slot}
\alias{pivot_lospat_slot}
\title{Pivot slot in list of single_period_at objects}
\usage{
pivot_lospat_slot(x, slotname, dates = NULL, simplify = TRUE)
}
\arguments{
\item{x}{list of single_period_at objects}

\item{slotname}{slot name to pivot in single_period_at object}

\item{dates}{an array of dates}

\item{simplify}{a logical. Is slot a value or list}
}
\value{
stackedorderedLists object
}
\description{
Function to pivot a list of orderedLists
 For use in converting a list of single_period_at objects to
 at slots.
}
