% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_z_score.R
\docType{class}
\name{return_z_score-class}
\alias{return_z_score-class}
\title{Return Z-Score (S4 Object)}
\description{
An S4 Class to represent Return Z-Scores
}
\section{Slots}{

\describe{
\item{\code{score}}{Return Z-Score}
}}

