% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-set_at_settings.R
\name{set_at_settings}
\alias{set_at_settings}
\title{Create Alpha Testing Settings}
\usage{
set_at_settings(
  testing_scheme = "factor-z",
  start_date = as.Date("1901-01-01"),
  end_date = Sys.Date(),
  weighting_scheme = NULL,
  benchmark_weighting_scheme = "zero",
  ...
)
}
\arguments{
\item{testing_scheme}{Weighting scheme to be used in testing factor.
Currently supports "Factor" and "Quantile"}

\item{start_date}{A date object representing the earliest start date to be
used when performing alpha testing. Factor data is filtered to start at or
after provided Start.Date argument.}

\item{end_date}{A date object representing the last date to be
used when performing alpha testing. Factor data is filtered to end at or
before provided End.Date argument.}

\item{weighting_scheme}{desctiption}

\item{benchmark_weighting_scheme}{description}

\item{...}{Additional settings to be passed based on Weighting.Scheme}
}
\description{
This function creates an "at_settings" class with arguments to be used when
testing factors
}
