% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_settings.R
\name{set_at_settings_factor_q}
\alias{set_at_settings_factor_q}
\title{Set Quantile Spread Alpha Testing Settings}
\usage{
set_at_settings_factor_q(
  start_date = as.Date("1901-01-01"),
  end_date = Sys.Date(),
  weighting_scheme = "equal",
  benchmark_weighting_scheme = "zero",
  quantiles = 5,
  ...
)
}
\arguments{
\item{start_date}{description}

\item{end_date}{description}

\item{weighting_scheme}{description}

\item{benchmark_weighting_scheme}{description}

\item{quantiles}{description}

\item{...}{Addtional arguements to be passesd to function.}
}
\value{
An at_settings_q_spread S4 object to be used in Alpha Testing.
}
\description{
A function to create Alpha Testing settings for a Quantile Spread Alpha Test.
}
