% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_settings.R
\name{set_at_settings_factor_z}
\alias{set_at_settings_factor_z}
\title{Set Factor Weighted Alpha Testing Settings}
\usage{
set_at_settings_factor_z(
  start_date = as.Date("1901-01-01"),
  end_date = Sys.Date(),
  weighting_scheme = "z-weighted",
  win_prob = c(0, 1),
  ...
)
}
\arguments{
\item{start_date}{description}

\item{end_date}{description}

\item{weighting_scheme}{description}

\item{win_prob}{A numeric vector of length 2 representing the percentile}

\item{...}{Additional arguements to be passed to function.}
}
\value{
An at_settings_factor_w S4 object to be used in Alpha Testing.
}
\description{
A function to create Alpha Testing settings for a factor weighted Alpha Test.
}
