% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_data.R
\name{set_date_col}
\alias{set_date_col}
\title{Set Date Column Name}
\usage{
set_date_col(x, date_col_name)
}
\arguments{
\item{x}{factor_data or factor_data_params object}

\item{date_col_name}{a string representing the date column name}
}
\value{
updated factor_data or factor_data_params object
}
\description{
Set the date column for factor_data_params
}
