% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_data.R
\name{set_factor_col}
\alias{set_factor_col}
\title{Set Factor Column Name}
\usage{
set_factor_col(x, factor_col_name)
}
\arguments{
\item{x}{factor_data or factor_data_params object}

\item{factor_col_name}{a string representing the factor column name}
}
\value{
updated factor_data or factor_data_params object
}
\description{
Set the Factor column for factor_data_params
}
