% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_data.R
\name{set_group_col}
\alias{set_group_col}
\title{Set Group Column Name}
\usage{
set_group_col(x, group_col_name)
}
\arguments{
\item{x}{factor_data or factor_data_params object}

\item{group_col_name}{a string representing the group column name}
}
\value{
updated factor_data or factor_data_params object
}
\description{
Set the Group column for factor_data_params
}
