% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_data.R
\name{set_id_col}
\alias{set_id_col}
\title{Set ID Column Name}
\usage{
set_id_col(x, date_col_name)
}
\arguments{
\item{x}{factor_data or factor_data_params object}

\item{id_col_name}{a string representing the id column name}
}
\value{
updated factor_data or factor_data_params object
}
\description{
Set the ID column for factor_data_params
}
