% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_data.R
\name{set_return_col}
\alias{set_return_col}
\title{Set Return Column Name}
\usage{
set_return_col(x, return_col_name)
}
\arguments{
\item{x}{factor_data or factor_data_params object}

\item{return_col_name}{a string representing the return column name}
}
\value{
updated factor_data or factor_data_params object
}
\description{
Set the Return column for factor_data_params
}
