% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_data.R
\name{set_weight_col}
\alias{set_weight_col}
\title{Set Weight Column Name}
\usage{
set_weight_col(x, weight_col_name)
}
\arguments{
\item{x}{factor_data or factor_data_params object}

\item{weight_col_name}{a string representing the weight column name}
}
\value{
updated factor_data or factor_data_params object
}
\description{
Set the Weight column for factor_data_params
}
