% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-single_period_at.R
\docType{class}
\name{single_period_at-class}
\alias{single_period_at-class}
\title{Single Period Alpha Testing S4 Object}
\description{
An S4 Class to represent Factor Alpha Testing Parent Class
 for a single period
}
\section{Slots}{

\describe{
\item{\code{date}}{A date object representing the date of the data}

\item{\code{return}}{portfolio return}

\item{\code{return_bmark}}{benchmark return}

\item{\code{weights}}{Portfolio Weights}

\item{\code{weights_bmark}}{benchmark weights}
}}

