% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-single_period_at_q.R
\docType{class}
\name{single_period_at_q-class}
\alias{single_period_at_q-class}
\title{Single Period Alpha Test (Quantiles) S4 Object}
\description{
Single Period Alpha Test (Quantiles) S4 Object
}
\section{Slots}{

\describe{
\item{\code{factor_quantile}}{todo}

\item{\code{q_returns}}{todo}

\item{\code{q_stats}}{todo}
}}

