% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-single_period_at_z.R
\docType{class}
\name{single_period_at_z-class}
\alias{single_period_at_z-class}
\title{Single Period Alpha Test (Factor Weighted) S4 Object}
\description{
Single Period Alpha Test (Factor Weighted) S4 Object
}
\section{Slots}{

\describe{
\item{\code{IC}}{Information Coefficient of Factor.}

\item{\code{factor_z_score}}{factor_z_score object.}

\item{\code{return_z_score}}{z-score of returns.}
}}

