% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePeriodAT.R
\docType{class}
\name{SinglePeriodAT-class}
\alias{SinglePeriodAT-class}
\title{An S4 Class to represent Factor Alpha Testing for a single period}
\description{
@slot .factordata A SinglePeriodFactorData object.
 @slot date A date object representing the date of the data
 @slot factorZscore A named numeric vector representing the factor Z-Score
 @slot factorQuantile A named numeric vector representing the factor Quantile
 @slot returnZscore A named numeric vector representing the forward
 return Z-Score
 @slot IC A numeric value representing the Information Coefficient
 @slot .settings
}
