% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePeriodAT.R
\name{SinglePeriodAT}
\alias{SinglePeriodAT}
\title{@title Single Period Factor Alpha Testing
 @description Function to create a Single Period Alpha Testing Object
 @param .data A SinglePeriodFactorData
 @param fftile A numeric value representing the number of quantiles to group the factor data by.
 @param win.prob A numeric vector of length 2 representing the percentile cut-offs for windsorization.
 @returns A SinglePeriodFactorData object.
 @keywords internal
 @export}
\usage{
SinglePeriodAT(.data, fftile = 5, win.prob = c(0, 1), ...)
}
\description{
@title Single Period Factor Alpha Testing
 @description Function to create a Single Period Alpha Testing Object
 @param .data A SinglePeriodFactorData
 @param fftile A numeric value representing the number of quantiles to group the factor data by.
 @param win.prob A numeric vector of length 2 representing the percentile cut-offs for windsorization.
 @returns A SinglePeriodFactorData object.
 @keywords internal
 @export
}
