% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePeriodAT.R
\docType{class}
\name{SinglePeriodAT_FactorWeighted-class}
\alias{SinglePeriodAT_FactorWeighted-class}
\title{Single Period Alpha Test (Factor Weighted) S4 Object}
\description{
Single Period Alpha Test (Factor Weighted) S4 Object
}
\section{Slots}{

\describe{
\item{\code{IC}}{todo}

\item{\code{factorZscore}}{todo}

\item{\code{returnZscore}}{todo}
}}

