% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePeriodAT.R
\docType{class}
\name{SinglePeriodAT_Quantile-class}
\alias{SinglePeriodAT_Quantile-class}
\title{Single Period Alpha Test (Quantiles) S4 Object}
\description{
Single Period Alpha Test (Quantiles) S4 Object
}
\section{Slots}{

\describe{
\item{\code{factorQuantile}}{todo}

\item{\code{qreturns}}{todo}

\item{\code{qstats}}{todo}
}}

