% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePeriodFactorData.R
\docType{class}
\name{SinglePeriodFactorData-class}
\alias{SinglePeriodFactorData-class}
\title{An S4 Class to represent Factor Data for a single period}
\description{
@slot factor A character representing the name of the factor
 @slot date A date object representing the date of the data
 @slot ids A character vector representing the company ids
 @slot fvales A named numeric vector representing the factor values
 @slot returns A names numeric vector representing the forward returns
}
