% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePeriodFactorData.R
\name{SinglePeriodFactorData}
\alias{SinglePeriodFactorData}
\title{Single Period Factor Data}
\usage{
SinglePeriodFactorData(data, date, iname, fname, rname, ...)
}
\arguments{
\item{data}{A data frame containing id columns, return column and factor
value column.}

\item{date}{A date object representing the date of the data.}

\item{iname}{A character representing the column name of identifiers.}

\item{fname}{A character representing the column name of the factor.}

\item{rname}{A character representing the column name of the returns.}
}
\value{
A SinglePeriodFactorData object.
}
\description{
Function to create SinglePeriodFactorData object
}
\details{
This function converts data into SinglePeriodFactorData object.
}
\keyword{internal}
