% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-stackorderedLists.R
\name{stackorderedLists}
\alias{stackorderedLists}
\title{Stack orderedLists}
\usage{
stackorderedLists(lol, y = NULL, simplify = TRUE)
}
\arguments{
\item{lol}{a list of orderedLists}

\item{y}{an array of names}

\item{simplify}{logical, should new data be combined or remain list}
}
\description{
Pivot the data in a list of orderedLists
}
