% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-alpha_test.R
\name{alpha_test}
\alias{alpha_test}
\title{Alpha Test Function}
\usage{
alpha_test(data, .settings, ...)
}
\arguments{
\item{data}{A `factor_data` or `single_period_factor_data` object containing
the input data for the alpha test.}

\item{.settings}{A `factor_data_params` object containing configuration
settings for the alpha test.}
}
\value{
A `multi_period_at` or `single_period_at` object, depending on the
  input data.
}
\description{
The `alpha_test` function performs a backtest on alpha test data using the
specified settings, generating performance metrics for factors.
}
\details{
The function computes backtest statistics, evaluating the efficacy of
factors in predicting excess returns. Metrics include forward cumulative
performance for the current factor value and one-period forward performance
for lagged factor values.
}
