% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_test.R, R/alpha_test_q.R
\docType{class}
\name{alpha_test_factor_q-class}
\alias{alpha_test_factor_q-class}
\title{Alpha Test (Factor Quantile) S4 Object}
\description{
Alpha Test (Factor Quantile) S4 Object

Alpha Test (Factor Quantile) S4 Object
}
\section{Slots}{

\describe{
\item{\code{factor_quantile}}{A list of factor quantiles.}

\item{\code{q_returns}}{A list of factor quantile returns.}

\item{\code{q_stats}}{A list of factor quantile stats.}

\item{\code{factor_quantile}}{A list of factor quantiles.}

\item{\code{q_returns}}{A list of factor quantile returns.}

\item{\code{q_stats}}{A list of factor quantile stats.}
}}

