% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePeriodAT.R
\name{AlphaTest}
\alias{AlphaTest}
\title{Single Period Factor Alpha Testing}
\usage{
AlphaTest(.Object)
}
\arguments{
\item{.data}{A SinglePeriodFactorData}

\item{fftile}{A numeric value representing the number of quantiles to group
the factor data by.}

\item{win.prob}{A numeric vector of length 2 representing the percentile
cut-offs for windsorization.}
}
\value{
A SinglePeriodFactorData object.
}
\description{
Function to create a Single Period Alpha Testing Object
}
\keyword{internal}
