% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_data.R
\docType{class}
\name{at_data-class}
\alias{at_data-class}
\title{Alpha Test Data (S4 Object)}
\description{
An S4 Class to represent alpha test data for multi-periods
}
\section{Slots}{

\describe{
\item{\code{alpha_data}}{A list of single_period_at_data objects.}

\item{\code{data}}{A data.frame containing raw data.}

\item{\code{params}}{An at_data_params object.}
}}

