% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-at_data_params.R
\docType{class}
\name{at_data_params-class}
\alias{at_data_params-class}
\title{Alpha Test Data Parameters}
\description{
An S4 class containing the parameters for configuring an `at_data`
object.
}
\section{Slots}{

\describe{
\item{\code{date_col_name}}{A character string representing the column name for
date values.}

\item{\code{id_col_name}}{A character string representing the column name for ID
values.}

\item{\code{factor_col_name}}{A character string representing the column name for
factor values.}

\item{\code{return_col_name}}{A character string representing the column name for
return values.}

\item{\code{include_col_name}}{A character string representing the column name for
inclusion flags (TRUE/FALSE).}

\item{\code{group_col_name}}{A character string representing the column name for
group values.}

\item{\code{weight_col_name}}{A character string representing the column name for
weights.}

\item{\code{horizon}}{A numeric value representing the return and factor horizon.}
}}

