% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-at_settings.R
\docType{class}
\name{at_settings-class}
\alias{at_settings-class}
\title{Alpha Testing Settings (S4 Object)}
\description{
An S4 Object containing the settings to be used in Alpha Testing. Arguments
passed through ATSettings are used in designing the alpha testing procedure.
}
\section{Slots}{

\describe{
\item{\code{start_date}}{Date representing the earliest date for Alpha Testing.}

\item{\code{end_date}}{Date representing the latest date for Alpha Testing.}

\item{\code{testing_scheme}}{Character representing alpha testing procedure.}

\item{\code{weighting_scheme}}{Character representing weighting scheme.}

\item{\code{benchmark_weighting_scheme}}{Character represenitng the weighting scheme
for the benchmark.}
}}

