% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_settings.R
\docType{class}
\name{at_settings_factor_q-class}
\alias{at_settings_factor_q-class}
\title{Settings for Factor Quantile Weighted Alpha Testing (S4 Object)}
\description{
Settings for Factor Quantile Weighted Alpha Testing (S4 Object)
}
\section{Slots}{

\describe{
\item{\code{quantiles}}{Number of Quantiles to use.}
}}

