% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-at_settings_factor_q.R
\docType{class}
\name{at_settings_factor_q-class}
\alias{at_settings_factor_q-class}
\title{Settings for Factor Quantile Weighted Alpha Testing (S4 Object)}
\description{
An S4 class for specifying settings used in quantile-based alpha testing.
}
\section{Slots}{

\describe{
\item{\code{quantiles}}{A numeric value representing the number of quantiles to use.}
}}

