% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-at_settings_factor_z.R
\docType{class}
\name{at_settings_factor_z-class}
\alias{at_settings_factor_z-class}
\title{Settings for Factor Z-Score Weighted Alpha Testing (S4 Object)}
\description{
An S4 class for specifying settings used in Z-score-based alpha testing.
}
\section{Slots}{

\describe{
\item{\code{win_prob}}{A numeric vector of length 2 specifying the percentile
cut-offs for winsorization.}
}}

