% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-calc_factor_q.R
\name{calc_factor_q,single_period_at_data-method}
\alias{calc_factor_q,single_period_at_data-method}
\title{Calculate Quantile of Factor Values}
\usage{
\S4method{calc_factor_q}{single_period_at_data}(at_data, quantiles = 5, .desc = TRUE)
}
\arguments{
\item{at_data}{A single_period_at_data S4 object.}

\item{quantiles}{a numeric representing the number of quantiles}
}
\value{
orderedList S4 object of factor_q_score S4 objects
}
\description{
Calculate the Quantile of Factor Values including lagged
 Factor Values in a single_period_at_data S4 object
}
