% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-calc_factor_q.R
\name{calc_factor_q}
\alias{calc_factor_q}
\title{Calculate Quantile Factor Value}
\usage{
calc_factor_q(at_data, quantiles, ...)
}
\arguments{
\item{at_data}{Alpha Test data}

\item{quantiles}{a numeric representing the number of quantiles}
}
\description{
Calculate Factor Quantile
}
