% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_q_score.R
\name{calc_factor_q}
\alias{calc_factor_q}
\title{Calculate Quantile Factor Value}
\usage{
calc_factor_q(factor_data, quantiles = 5)
}
\arguments{
\item{quantiles}{a numeric representing the number of quantiles.}

\item{factor_values}{a numeric array of factor values}
}
\value{
Ordered quantiles
}
\description{
Calculate Factor Z-Score
}
