% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-calc_factor_z.R
\name{calc_factor_z,single_period_at_data-method}
\alias{calc_factor_z,single_period_at_data-method}
\title{Calculate Z-Score of Factor Values}
\usage{
\S4method{calc_factor_z}{single_period_at_data}(at_data, win_prob = c(0, 1))
}
\arguments{
\item{at_data}{A single_period_at_data S4 object.}

\item{win_prob}{a numeric vector of probabilities with values in [0,1]
defining the windsorization levels.}
}
\value{
orderedList S4 object of factor_z_score S4 objects
}
\description{
Calculate the Z-Score of Factor Values including lagged
 Factor Values in a single_period_at_data S4 object
}
