% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_return.R
\name{calc_return}
\alias{calc_return}
\title{Calculate Weighted Return from Weights & Returns}
\usage{
calc_return(weights, returns, ...)
}
\arguments{
\item{weights}{An array of weights or orderedList of weights}

\item{returns}{An array of returns or orderedList of returns
used to calculate the portfolio weights.}
}
\description{
Function to calculate the portfolio return from Weights
 and Returns
}
