% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-calc_return_z.R
\name{calc_return_z,single_period_at_data-method}
\alias{calc_return_z,single_period_at_data-method}
\title{Calculate Z-Score of Returns in Single Period AT Data}
\usage{
\S4method{calc_return_z}{single_period_at_data}(at_data, win_prob = c(0, 1))
}
\value{
orderedList S4 object of return_z_score S4 objects
}
\description{
Calculate Z-Score of Returns in Single Period AT Data
}
