% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_z_score.R
\name{calc_return_z}
\alias{calc_return_z}
\title{Calculate Z-Score of Returns}
\usage{
calc_return_z(return_data, win_prob = c(0, 1))
}
\arguments{
\item{return_data}{Returns}

\item{win_prob}{a numeric vector of probabilities with values in [0,1]}
}
\description{
Calculate Factor Z-Score
}
