% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_data_params.R
\name{create_at_data_params}
\alias{create_at_data_params}
\title{Create Alpha Test Data Parameters Object}
\usage{
create_at_data_params(
  date_col_name = NA_character_,
  id_col_name = NA_character_,
  factor_col_name = NA_character_,
  return_col_name = NA_character_,
  include_col_name = NA_character_,
  group_col_name = NA_character_,
  weight_col_name = NA_character_,
  horizon = 12
)
}
\arguments{
\item{date_col_name}{A string representing the column name of date values.}

\item{id_col_name}{A string representing the column name of id values.}

\item{factor_col_name}{A string representing the column name of factor values. # nolint: line_length_linter.}

\item{return_col_name}{A string representing the column name of return values. # nolint: line_length_linter.}

\item{include_col_name}{A string representing the column name of include T/F.}

\item{group_col_name}{A string representing the column name of grouping value. # nolint: line_length_linter.}

\item{weight_col_name}{A string representing the column name of weights.}

\item{horizon}{A numeric representing the return and factor horizon.}
}
\value{
at_data_params object
}
\description{
Function to set parameters for factor data, creating a at_data_params object.
}
