% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-create_factor_data_params.R
\name{create_factor_data_params}
\alias{create_factor_data_params}
\title{Create Factor Data Parameters Object}
\usage{
create_factor_data_params(
  date_col_name = NA_character_,
  id_col_name = NA_character_,
  factor_col_name = NA_character_,
  return_col_name = NA_character_,
  group_col_name = NA_character_,
  weight_col_name = NA_character_
)
}
\arguments{
\item{date_col_name}{A string representing the column name of date values.}

\item{id_col_name}{A string representing the column name of id values.}

\item{factor_col_name}{A string representing the column name of factor values.}

\item{return_col_name}{A string representing the column name of return values.}

\item{group_col_name}{A string representing the column name of grouping value.}

\item{weight_col_name}{A string representing the column name of weights.}
}
\value{
factor_data_params object
}
\description{
Function to set parameters for factor data, creating a factor_data_params
object.
}
