% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-create_single_period_at_data.R
\name{create_single_period_at_data}
\alias{create_single_period_at_data}
\title{Create Single Period Alpha Test Data}
\usage{
create_single_period_at_data(
  date,
  data,
  date_col_name,
  id_col_name,
  factor_col_name,
  return_col_name,
  include_col_name,
  group_col_name,
  weight_col_name,
  horizon = 0
)
}
\arguments{
\item{date}{A `Date` object representing the specific period.}

\item{data}{A data frame containing the required columns: ID, return, and
factor values.}

\item{date_col_name}{A character string for the column name of dates.}

\item{id_col_name}{A character string for the column name of IDs.}

\item{factor_col_name}{A character string for the column name of factors.}

\item{return_col_name}{A character string for the column name of returns.}

\item{include_col_name}{A character string for the column name of include
flags (TRUE/FALSE).}

\item{group_col_name}{A character string for the column name of groups.}

\item{weight_col_name}{A character string for the column name of weights.}

\item{horizon}{A numeric value representing the horizon for returns and
factors. Defaults to 0.}
}
\value{
A `single_period_at_data` S4 object.
}
\description{
Converts raw data into a `single_period_at_data` object for single-period
alpha testing.
}
\keyword{internal}
