% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_q_score.R
\docType{class}
\name{factor_q_score-class}
\alias{factor_q_score-class}
\title{Factor Quantile (S4 Object)}
\description{
An S4 Class to represent Factor Quantiles
}
\section{Slots}{

\describe{
\item{\code{score}}{Factor Quantile}
}}

