% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_z_score.R
\docType{class}
\name{factor_z_score-class}
\alias{factor_z_score-class}
\title{Factor Z-Score (S4 Object)}
\description{
An S4 Class to represent Factor Z-Scores
}
\section{Slots}{

\describe{
\item{\code{score}}{Factor Z Scores}
}}

