% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_fwd_p_returns}
\alias{get_fwd_p_returns}
\title{Get forward periodic return values}
\usage{
get_fwd_p_returns(data, date_col_name, id_col_name, return_col_name, horizon)
}
\arguments{
\item{data}{Data frame}

\item{date_col_name}{string representing name of date column in data.}

\item{id_col_name}{string representing name of id column in data.}

\item{return_col_name}{string representing name of return in data.}

\item{horizon}{numeric representing the number of forward return periods.}
}
\description{
For every date, id get the periodic forward returns.
}
