% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_prv_factor_vals}
\alias{get_prv_factor_vals}
\title{Compute Lagged Factor Values}
\usage{
get_prv_factor_vals(data, date_col_name, id_col_name, factor_col_name, horizon)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{date_col_name}{A string specifying the column name for dates.}

\item{id_col_name}{A string specifying the column name for IDs.}

\item{factor_col_name}{A string specifying the column name for factors.}

\item{horizon}{An integer representing the lag horizon in periods.}
}
\value{
A data frame containing lagged factor values.
}
\description{
Computes lagged factor values for a given dataset across specified time
horizons.
}
