% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-multi_period_at_q.R
\docType{class}
\name{multi_period_at_q-class}
\alias{multi_period_at_q-class}
\title{Multi-Period Alpha Test Data (Quantile-Based) (S4 Object)}
\description{
Encapsulates multi-period alpha test results for quantile-based analysis.
}
\section{Slots}{

\describe{
\item{\code{factor_quantile}}{An `orderedList` of factor quantiles for each period.}

\item{\code{q_avg_fwd_returns}}{An `orderedList` of average forward returns by
quantile.}

\item{\code{q_lag_factor_avg_return}}{An `orderedList` of average returns for lagged
factors by quantile.}
}}

