% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-multi_period_at_z.R
\docType{class}
\name{multi_period_at_z-class}
\alias{multi_period_at_z-class}
\title{Multi-Period Alpha Test Data (Z-Score-Based) (S4 Object)}
\description{
Represents multi-period alpha test data for Z-score-based analysis.
}
\section{Slots}{

\describe{
\item{\code{IC_fwd_return}}{An `orderedList` of information coefficients for
forward returns.}

\item{\code{IC_lag_factor}}{An `orderedList` of information coefficients for lagged
factors.}
}}

