% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-orderedList.R
\docType{class}
\name{orderedList-class}
\alias{orderedList-class}
\title{orderedList (S4 Object)}
\description{
An S4 Class to represent a list with a given order
}
\section{Slots}{

\describe{
\item{\code{list}}{A list}

\item{\code{n}}{the length of the list}

\item{\code{order}}{a numeric array representing the order of the list}
}}

