% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-return_z_score.R
\docType{class}
\name{return_z_score-class}
\alias{return_z_score-class}
\title{Return Z-Score (S4 Object)}
\description{
Represents Z-scores of returns for a single period in alpha testing.
}
\section{Slots}{

\describe{
\item{\code{score}}{A numeric vector containing Z-scores for returns.}
}}

