% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-single_period_at_data.R
\docType{class}
\name{single_period_at_data-class}
\alias{single_period_at_data-class}
\title{Single Period Alpha Test Data (S4 Object)}
\description{
An S4 Class to represent Alpha Test Data for a single period.
}
\section{Slots}{

\describe{
\item{\code{factor}}{A character representing the name of the factor.}

\item{\code{date}}{A date object representing the date of the data.}

\item{\code{ids}}{A character vector representing the company ids.}

\item{\code{fvales}}{An orderedList of factor values.}

\item{\code{returns}}{An orderedList of returns.}

\item{\code{group}}{A named character vector representing the company grouping.}

\item{\code{weights}}{A named numeric vector of weights from factor values.}
}}

