% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePeriodAT.R
\docType{class}
\name{SinglePeriodAT-class}
\alias{SinglePeriodAT-class}
\title{An S4 Class to represent Factor Alpha Testing for a single period}
\description{
An S4 Class to represent Factor Alpha Testing for a single period
}
\section{Slots}{

\describe{
\item{\code{.factordata}}{A SinglePeriodFactorData object.}

\item{\code{date}}{A date object representing the date of the data}

\item{\code{factorZscore}}{A named numeric vector representing the factor Z-Score}

\item{\code{factorQuantile}}{A named numeric vector representing the factor Quantile}

\item{\code{returnZscore}}{A named numeric vector representing the forward
return Z-Score}

\item{\code{IC}}{A numeric value representing the Information Coefficient}

\item{\code{uStats}}{Universe Return Statistics}

\item{\code{qStats}}{Quantile Return Statistics}

\item{\code{.settings}}{Alpha Testing settings}
}}

