% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePeriodAT.R
\docType{class}
\name{SinglePeriodAT-class}
\alias{SinglePeriodAT-class}
\title{Single Period Alpha Testing S4 Object}
\description{
An S4 Class to represent Factor Alpha Testing Parent Class
 for a single period
}
\section{Slots}{

\describe{
\item{\code{date}}{A date object representing the date of the data}

\item{\code{return}}{}

\item{\code{alpha}}{Relative Return of Factor over Benchmark}

\item{\code{weights}}{Portfolio Weights}

\item{\code{.factordata}}{A SinglePeriodFactorData object.#'}

\item{\code{.settings}}{Alpha Testing Settings#'}
}}

