% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePeriodFactorData.R
\docType{class}
\name{SinglePeriodFactorData-class}
\alias{SinglePeriodFactorData-class}
\title{An S4 Class to represent Factor Data for a single period}
\description{
An S4 Class to represent Factor Data for a single period
}
\section{Slots}{

\describe{
\item{\code{factor}}{A character representing the name of the factor}

\item{\code{date}}{A date object representing the date of the data}

\item{\code{ids}}{A character vector representing the company ids}

\item{\code{fvales}}{A named numeric vector representing the factor values}

\item{\code{returns}}{A names numeric vector representing the forward returns}
}}

